requirements.txt
# Core dependencies
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
pydantic>=2.5.0
pydantic-settings>=2.1.0

# Data processing
pandas>=2.0.0
numpy>=1.24.0
duckdb>=0.9.0

# HuggingFace datasets
datasets>=2.14.0
huggingface-hub>=0.16.0

# Configuration
pyyaml>=6.0

# Utilities
python-dotenv>=1.0.0
tqdm>=4.65.0

# Testing (optional)
pytest>=7.4.0
pytest-asyncio>=0.21.0

